requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pyttsx3>=2.90
pydantic>=2.8.0